tests/wpt/web-platform-tests/html/interaction/focus/focusgroup/tentative/backward-navigation/does-not-move-when-on-focusgroup-root.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: focusgroup - Focus does not move when initially set on the focusgroup root.</title>
<link rel="author" title="Microsoft" href="http://www.microsoft.com/">
<link rel="help" href="https://github.com/MicrosoftEdge/MSEdgeExplainers/blob/main/Focusgroup/explainer.md">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="../resources/focusgroup-utils.js"></script>

<div id=root tabindex=-1 focusgroup>
  <span id=item1 tabindex=0>item1</span>
  <span id=item2 tabindex=-1>item2</span>
</div>

<script>

  promise_test(async t => {
    var root = document.getElementById("root");

    await focusAndKeyPress(root, kArrowUp);
    assert_equals(document.activeElement, root);

    await focusAndKeyPress(root, kArrowLeft);
    assert_equals(document.activeElement, root);
  }, "When the focus is set on the root of a focusgroup element, an arrow key press shouldn't move the focus at all.");

</script>
